requirements.txt
# Copy
python-dotenv
fastapi
uvicorn
google-generativeai
pillow
jinja2